flowedit: Rename canvas_view.rs to flow_canvas.rs (#2611)
Rename module to better reflect its primary purpose of housing
the FlowCanvas struct and related rendering code.
